Move map_elements to separate module
